+++ yahtzee_api/game.py
@@ -14,6 +14,7 @@ class Game:
         remaining_turns (int): Global turn tracker.
         current_player (Player): The player who is currently rolling dice/scoring.
         num_players (int): Number of players in the game. 
+        winner (list): list populated at the end of the game with Player object(s) to store winner(s) (in case of a tie)
     """
 
     def __init__(self, num_players):
@@ -26,45 +27,72 @@ class Game:
         self.remaining_turns = 13
         self.current_player = self.players[0]
         self.num_players = num_players
+        self.winner = []
 
     def next_player(self):
         """Advances to the next player and moves to the next global turn when the last player is detected.
         
         Once the 13th turn is completed, this method will automatically calculate the final scores and
-        return a list containing the winner(s) Player object(s).
-
-        Returns:
-            Either an empty list if the game is still in progress,
-            or a list of the Player objects of the winner(s) 
-            (multiple winners in the case of a tie).
+        set the value of self.winner to the winner(s) Player object(s).
         """
         if self.players.index(self.current_player) == self.num_players - 1:
             self.remaining_turns -= 1
             self.current_player = self.players[0]
             if self.remaining_turns == 0:
-                return self._end_game()
+                self._end_game()
         else:
             self.current_player = self.players[self.players.index(self.current_player) + 1]
-            return []
 
-    def print_status(self):
-        """Prints out the current moment-in-time status of the game."""
-        print("Turns Remaining: ", self.remaining_turns)
-        print("Current player:", self.current_player.player_name)
-        print("\n")
-        for player in self.players:
-            print(player.player_name + "'s scorecard:")
-            player.print_scorecard()
+    def print_status(self, file, overwrite=True):
+        """Prints out the current moment-in-time status of the game to a specified file.
+        
+        Args: 
+            file (str): Filename to write to.
+            overwrite (bool, optional): Whether or not to overwrite existing file data. Defaults to True.
+        """
+        with open(file, 'w' if overwrite == True else 'a') as f:
+            f.write("Current Turn: " + str(13 - self.remaining_turns + 1) + "\n")
+            f.write("Turns Remaining: " + str(self.remaining_turns - 1) + "\n")
+            f.write("Current Player: " + self.current_player.player_name + "\n")
+            f.write("Current dice: " + str(self.current_player.dice) + "\n")
+            f.write("Rolls left this turn: " + str(self.current_player.rolls_left) + "\n")
+            f.write("Scorecard:" + "\n")
+            for i in range(13):
+                f.write(str(self.current_player.scorecard[i]) + "\n")
+            f.write("\n")
+            f.write("Theoretical Scorecard:" + "\n")
+            for i in range(13):
+                f.write(str(self.current_player.theoretical_scorecard[i]) + "\n")
+            f.write("\n")
+            f.write("-----------------------------------------")
+            f.write("\n")
+        f.close()
+
+    def print_final(self, file, overwrite=True):
+        """Prints out the final results of the game to a specified file.
+        
+        Args: 
+            file (str): Filename to write to.
+            overwrite (bool, optional): Whether or not to overwrite existing file data. Defaults to True.
+        """
+        with open(file, 'w' if overwrite == True else 'a') as f:
+            f.write("Winner(s): ")
+            for player in self.winner:
+                f.write(player.player_name + "\n")
+                f.write("Scorecard:" + "\n")
+                for i in range(13):
+                    f.write(str(self.current_player.scorecard[i]) + "\n")
+                f.write("\n")
+            f.write("Score: " + str(self.winner[0].score) + "\n")
+            f.write("-----------------------------------------")
+            f.write("\n")
+        f.close()
 
     def _end_game(self):
-        """Computes final scores for each player and prints the results.
-        
-        Returns a list with the Player objects of the winner(s). If there is a tie, the list will contain all players who had the highest score.
-        Otherwise the list will contain only the winner, in position 0.
-        """
+        """Computes final scores for each player and prints the results."""
         final_scores = []
         for player in self.players:
             player.calculate_final_score()
             final_scores.append(player.score)
-        return [player for player in self.players if player.score == max(final_scores)]
+        self.winner = [player for player in self.players if player.score == max(final_scores)]